% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_results.R
\name{chess_results}
\alias{chess_results}
\title{Get all Data for a Tournament}
\usage{
chess_results(id)
}
\arguments{
\item{id}{The URL of the chess-results.com tournament page or
the id that is present in the URL, supplied as a string.}
}
\value{
A list of tibbles about the different aspects of the tournament.
}
\description{
For a given URL, or tournament ID (which can be found in the URL),
this function scraps a list of tibbles and related data which are
found on that page.
}
\examples{
chess_results("1443765")

chess_results("https://chess-results.com/tnr134.aspx")
}
