% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_results.R
\name{chess_results}
\alias{chess_results}
\title{Get all data for a tournament}
\usage{
chess_results(id, user_agent = "chessResults R package")
}
\arguments{
\item{id}{The URL or the tournament ID of the tournament page}

\item{user_agent}{The User Agent to be used when accessing
chess-results.com}
}
\value{
A list of tibble about the different aspects of
the tournament. The sub-list contains the data for
the pairings/results for each round.
}
\description{
Scrape a list of tibble from chess-results.com by giving a URL
or tournament ID. (The tournament ID can be found in the URL).
It currently returns tournament information, starting rank,
playing schedule, pairings/results for each round, and closing rank.
}
\examples{
\donttest{
chess_results("https://s3.chess-results.com/tnr1445162.aspx?lan=1")
}
}
